pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bangen"
version = "2.2.0"
description = "ASCII banner rendering engine with a tiered effect library, live TUI, and transparent PNG/GIF export."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "programmersd21" }]
keywords = ["ascii", "banner", "terminal", "pyfiglet", "rich", "gradient", "tui"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Utilities",
  "Topic :: Terminals",
]
dependencies = [
  "pyfiglet>=1.0",
  "rich>=13.0",
  "Pillow>=10.0",
]

[project.optional-dependencies]
images = []
all = []

[project.urls]
Homepage   = "https://github.com/programmersd21/bangen"
Repository = "https://github.com/programmersd21/bangen"
Issues     = "https://github.com/programmersd21/bangen/issues"
Changelog  = "https://github.com/programmersd21/bangen/blob/main/CHANGELOG.md"

[project.scripts]
bangen = "bangen.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["bangen*"]

[tool.setuptools]
py-modules = ["main"]
